feat: simplify and rename test/send_test.dart
